gstreamer_video::caps

Struct VideoCapsBuilder

pub struct VideoCapsBuilder<T> {
    builder: Builder<T>,
}

Fields
builder: Builder<T>
Implementations

impl VideoCapsBuilder<NoFeature>

pub fn new() -> Self

Constructs an VideoCapsBuilder for the “video/x-raw” encoding.

If left unchanged, the resulting Caps will be initialized with:

“video/x-raw” encoding.

all available formats.

maximum width range.

maximum height range.

Use VideoCapsBuilder::for_encoding to specify another encoding.

pub fn for_encoding(encoding: impl IntoGStr) -> Self

Constructs an VideoCapsBuilder for the specified encoding.

The resulting Caps will use the encoding argument as name and will not contain any additional fields unless explicitly added.

pub fn for_encoding_from_static(encoding: impl AsRef<GStr> + 'static) -> Self

Constructs an VideoCapsBuilder for the specified encoding.

The resulting Caps will use the encoding argument as name and will not contain any additional fields unless explicitly added.

pub fn for_encoding_from_id(encoding: impl AsRef<IdStr>) -> Self

Constructs an VideoCapsBuilder for the specified encoding.

The resulting Caps will use the encoding argument as name and will not contain any additional fields unless explicitly added.

pub fn any_features(self) -> VideoCapsBuilder<HasFeatures>

pub fn features<S: IntoGStr>( self, features: impl IntoIterator<Item = S>, ) -> VideoCapsBuilder<HasFeatures>

pub fn features_from_statics<S: AsRef<GStr> + 'static>( self, features: impl IntoIterator<Item = S>, ) -> VideoCapsBuilder<HasFeatures>

pub fn features_from_ids<S: AsRef<IdStr>>( self, features: impl IntoIterator<Item = S>, ) -> VideoCapsBuilder<HasFeatures>

impl<T> VideoCapsBuilder<T>

pub fn format(self, format: VideoFormat) -> Self

pub fn format_if(self, format: VideoFormat, predicate: bool) -> Self

pub fn format_if_some(self, format: Option<VideoFormat>) -> Self

pub fn format_list(self, formats: impl IntoIterator<Item = VideoFormat>) -> Self

pub fn format_list_if( self, formats: impl IntoIterator<Item = VideoFormat>, predicate: bool, ) -> Self

pub fn format_list_if_some( self, formats: Option<impl IntoIterator<Item = VideoFormat>>, ) -> Self

pub fn format_list_if_not_empty( self, formats: impl IntoIterator<Item = VideoFormat>, ) -> Self

pub fn width(self, width: i32) -> Self

pub fn width_if(self, width: i32, predicate: bool) -> Self

pub fn width_if_some(self, width: Option<i32>) -> Self

pub fn width_range(self, widths: impl RangeBounds<i32>) -> Self

pub fn width_range_if_some(self, widths: Option<impl RangeBounds<i32>>) -> Self

pub fn width_range_if( self, widths: impl RangeBounds<i32>, predicate: bool, ) -> Self

pub fn width_list(self, widths: impl IntoIterator<Item = i32>) -> Self

pub fn width_list_if( self, widths: impl IntoIterator<Item = i32>, predicate: bool, ) -> Self

pub fn width_list_if_some( self, widths: Option<impl IntoIterator<Item = i32>>, ) -> Self

pub fn width_list_if_not_empty( self, widths: impl IntoIterator<Item = i32>, ) -> Self

pub fn height(self, height: i32) -> Self

pub fn height_if(self, height: i32, predicate: bool) -> Self

pub fn height_if_some(self, height: Option<i32>) -> Self

pub fn height_range(self, heights: impl RangeBounds<i32>) -> Self

pub fn height_range_if( self, heights: impl RangeBounds<i32>, predicate: bool, ) -> Self

pub fn height_range_if_some( self, heights: Option<impl RangeBounds<i32>>, ) -> Self

pub fn height_list(self, heights: impl IntoIterator<Item = i32>) -> Self

pub fn height_list_if( self, heights: impl IntoIterator<Item = i32>, predicate: bool, ) -> Self

pub fn height_list_if_some( self, heights: Option<impl IntoIterator<Item = i32>>, ) -> Self

pub fn height_list_if_not_empty( self, heights: impl IntoIterator<Item = i32>, ) -> Self

pub fn framerate(self, framerate: Fraction) -> Self

pub fn framerate_if(self, framerate: Fraction, predicate: bool) -> Self

pub fn framerate_if_some(self, framerate: Option<Fraction>) -> Self

pub fn framerate_range(self, framerates: impl RangeBounds<Fraction>) -> Self

pub fn framerate_range_if( self, framerates: impl RangeBounds<Fraction>, predicate: bool, ) -> Self

pub fn framerate_range_if_some( self, framerates: Option<impl RangeBounds<Fraction>>, ) -> Self

pub fn framerate_list( self, framerates: impl IntoIterator<Item = Fraction>, ) -> Self

pub fn framerate_list_if( self, framerates: impl IntoIterator<Item = Fraction>, predicate: bool, ) -> Self

pub fn framerate_list_if_some( self, framerates: Option<impl IntoIterator<Item = Fraction>>, ) -> Self

pub fn framerate_list_if_not_empty( self, framerates: impl IntoIterator<Item = Fraction>, ) -> Self

pub fn pixel_aspect_ratio(self, pixel_aspect_ratio: Fraction) -> Self

pub fn pixel_aspect_ratio_if( self, pixel_aspect_ratio: Fraction, predicate: bool, ) -> Self

pub fn pixel_aspect_ratio_if_some( self, pixel_aspect_ratio: Option<Fraction>, ) -> Self

pub fn pixel_aspect_ratio_range( self, pixel_aspect_ratios: impl RangeBounds<Fraction>, ) -> Self

pub fn pixel_aspect_ratio_range_if( self, pixel_aspect_ratios: impl RangeBounds<Fraction>, predicate: bool, ) -> Self

pub fn pixel_aspect_ratio_range_if_some( self, pixel_aspect_ratios: Option<impl RangeBounds<Fraction>>, ) -> Self

pub fn pixel_aspect_ratio_list( self, pixel_aspect_ratios: impl IntoIterator<Item = Fraction>, ) -> Self

pub fn pixel_aspect_ratio_list_if( self, pixel_aspect_ratios: impl IntoIterator<Item = Fraction>, predicate: bool, ) -> Self

pub fn pixel_aspect_ratio_list_if_some( self, pixel_aspect_ratios: Option<impl IntoIterator<Item = Fraction>>, ) -> Self

pub fn pixel_aspect_ratio_list_if_not_empty( self, pixel_aspect_ratios: impl IntoIterator<Item = Fraction>, ) -> Self

pub fn field(self, name: impl IntoGStr, value: impl Into<Value> + Send) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_with_static( self, name: impl AsRef<GStr> + 'static, value: impl Into<Value> + Send, ) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_with_id( self, name: impl AsRef<IdStr>, value: impl Into<Value> + Send, ) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_if( self, name: impl IntoGStr, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_with_static_if( self, name: impl AsRef<GStr> + 'static, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_with_id_if( self, name: impl AsRef<IdStr>, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_if_some( self, name: impl IntoGStr, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_with_static_if_some( self, name: impl AsRef<GStr> + 'static, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_with_id_if_some( self, name: impl AsRef<IdStr>, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_with_static_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_with_id_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn field_with_static_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn field_with_id_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn build(self) -> Caps