Trait gstreamer_video::prelude::BufferPoolExt

pub trait BufferPoolExt {
    fn get_options(&self) -> Vec<String>;

    fn has_option(&self, option: &str) -> bool;

    fn is_active(&self) -> bool;

    fn set_active(&self, active: bool) -> Result<(), BoolError>;

    fn set_flushing(&self, flushing: bool);
}

Trait containing all BufferPool methods.

Implementors

BufferPool

Required Methods

Important traits for Vec<u8>
impl Write for Vec<u8>

fn get_options(&self) -> Vec<String>

Get a None terminated array of string with supported bufferpool options for self. An option would typically be enabled with BufferPool::config_add_option.

Returns

a None terminated array of strings.

fn has_option(&self, option: &str) -> bool

Check if the bufferpool supports option.

option

an option

Returns

true if the buffer pool contains option.

fn is_active(&self) -> bool

Check if self is active. A pool can be activated with the BufferPoolExt::set_active call.

Returns

true when the pool is active.

fn set_active(&self, active: bool) -> Result<(), BoolError>

Control the active state of self. When the pool is inactive, new calls to BufferPool::acquire_buffer will return with FlowReturn::Flushing.

Activating the bufferpool will preallocate all resources in the pool based on the configuration of the pool.

Deactivating will free the resources again when there are no outstanding buffers. When there are outstanding buffers, they will be freed as soon as they are all returned to the pool.

active

the new active state

Returns

false when the pool was not configured or when preallocation of the buffers failed.

fn set_flushing(&self, flushing: bool)

Enable or disable the flushing state of a self without freeing or allocating buffers.

flushing

whether to start or stop flushing