Struct gstreamer_video::VideoInfo

pub struct VideoInfo(_);

Information describing image properties. This information can be filled in from GstCaps with VideoInfo::from_caps. The information is also used to store the specific video info when mapping a video frame with VideoFrame::map.

Use the provided macros to access the info in this structure.

Methods

impl VideoInfo

fn new<'a>(format: VideoFormat, width: u32, height: u32) -> VideoInfoBuilder<'a>

Allocate a new VideoInfo that is also initialized with VideoInfo::init.

Returns

a new VideoInfo. free with VideoInfo::free.

fn from_caps(caps: &Caps) -> Option<Self>

fn to_caps(&self) -> Option<Caps>

fn format(&self) -> VideoFormat

fn format_info(&self) -> VideoFormatInfo

fn width(&self) -> u32

fn height(&self) -> u32

fn interlace_mode(&self) -> VideoInterlaceMode

fn flags(&self) -> VideoFlags

fn size(&self) -> usize

fn views(&self) -> u32

fn chroma_site(&self) -> VideoChromaSite

fn colorimetry(&self) -> VideoColorimetry

fn par(&self) -> Fraction

fn fps(&self) -> Fraction

fn offset(&self) -> &[usize]

fn stride(&self) -> &[i32]

fn multiview_mode(&self) -> VideoMultiviewMode

fn multiview_flags(&self) -> VideoMultiviewFlags

fn field_order(&self) -> VideoFieldOrder

fn has_alpha(&self) -> bool

fn is_gray(&self) -> bool

fn is_rgb(&self) -> bool

fn is_yuv(&self) -> bool

fn is_interlaced(&self) -> bool

fn n_planes(&self) -> u32

fn n_components(&self) -> u32

fn convert(
&self,
src_fmt: Format,
src_val: i64,
dest_fmt: Format
) -> Option<i64>