Struct gstreamer_video::VideoInfo

pub struct VideoInfo(_);

Information describing image properties. This information can be filled in from GstCaps with VideoInfo::from_caps. The information is also used to store the specific video info when mapping a video frame with VideoFrame::map.

Use the provided macros to access the info in this structure.

Methods

impl VideoInfo

pub fn new<'a>(
format: VideoFormat,
width: u32,
height: u32
) -> VideoInfoBuilder<'a>

Allocate a new VideoInfo that is also initialized with VideoInfo::init.

Returns

a new VideoInfo. free with VideoInfo::free.

pub fn from_caps(caps: &Caps) -> Option<Self>

pub fn to_caps(&self) -> Option<Caps>

pub fn format(&self) -> VideoFormat

pub fn format_info(&self) -> VideoFormatInfo

pub fn width(&self) -> u32

pub fn height(&self) -> u32

pub fn interlace_mode(&self) -> VideoInterlaceMode

pub fn flags(&self) -> VideoFlags

pub fn size(&self) -> usize

pub fn views(&self) -> u32

pub fn chroma_site(&self) -> VideoChromaSite

pub fn colorimetry(&self) -> VideoColorimetry

pub fn par(&self) -> Fraction

pub fn fps(&self) -> Fraction

pub fn offset(&self) -> &[usize]

pub fn stride(&self) -> &[i32]

pub fn multiview_mode(&self) -> VideoMultiviewMode

pub fn multiview_flags(&self) -> VideoMultiviewFlags

pub fn field_order(&self) -> VideoFieldOrder

pub fn has_alpha(&self) -> bool

pub fn is_gray(&self) -> bool

pub fn is_rgb(&self) -> bool

pub fn is_yuv(&self) -> bool

pub fn is_interlaced(&self) -> bool

pub fn n_planes(&self) -> u32

pub fn n_components(&self) -> u32

pub fn convert<V: Into<FormatValue>>(
&self,
src_val: V,
dest_fmt: Format
) -> Option<FormatValue>